requirements.txt
cycler==0.10.0
decorator==4.4.0
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
opencv-contrib-python==4.1.1.26
Pillow==6.2.0
pkg-resources==0.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.16.1
scipy==1.3.1
six==1.12.0
